requirements.txt
streamlit
fastapi
uvicorn
langchain
langchain-community
langchain-huggingface
sentence-transformers
faiss-cpu
pypdf
python-docx
python-dotenv
